refactor: the flag set fills the whole config
newFlagSet took a second out-parameter for -color, because flag has no
ColorMode. That leak reached both call sites: printUsage declared a colour
variable it filled and never read, and parseFlags carried one plus a branch to
convert it after parsing. Meanwhile -fail-under, with the harder type problem,
already parsed itself.

-color parses itself the same way now, so newFlagSet takes one argument and a
parsed set is a finished config. ColorAuto is the zero value, so the default
falls out without being stated.

That also cleaned up the error, which the flag package now prefixes: it read
"invalid value \"sometimes\" for flag -color: invalid -color, want ..., got
\"sometimes\"". The sentinel is terse, like the one beside it.

Rendering help and parsing arguments shared a file only because both touch a
FlagSet. usage.go holds the text and printUsage.
